refactor email sending (#59)
Main reason for this being to provide the ability for projects to completely overwrite ``DefaultEmailSender`` and run their own logic when it comes to sending out emails (e.g. in my case through a third party service with very specific requirements).

At the same time this decouples the email sending from forms, views and models nicely.